requirements.txt
Flask==2.2.5
requests==2.28.1
google-auth==2.16.0
google-api-python-client==2.70.0
langchain==0.0.164
faiss-cpu==1.7.3
transformers
sentence-transformers
PyPDF2
pypdf
flask_cors